@@ -1,13 +1,8 @@
-
-# -*- coding: utf-8 -*-
-
-__all__ = [
-    'OnToma'
-]
-from ontoma.interface import OnToma
-
+__all__ = ['OnToma']
 
 import logging
+
+from ontoma.interface import OnToma
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -15,6 +10,3 @@ ch = logging.StreamHandler()
 formatter = logging.Formatter('%(levelname)-8s - %(name)s - %(message)s')
 ch.setFormatter(formatter)
 logger.addHandler(ch)
-
-
-
